Refactored sessionId:int to sessionId:long

+++ src/v20170926/runTest/SessionTest.java
@@ -8,8 +8,8 @@ public class SessionTest {
 
     public static void main(String[] args) {
 
-        int sessionId1 = sessionList.newSession();
-        int sessionId2 = sessionList.newSession();
+        long sessionId1 = sessionList.newSession();
+        long sessionId2 = sessionList.newSession();
 
         System.out.println("sessionId1 = "+sessionId1);
         System.out.println("sessionId2 = "+sessionId2);
